Move middleware.socket
